+++ resources/views/partials/fields/horizontal.blade.php
@@ -1,6 +1,6 @@
 <div class="form-group row">
     @isset($title)
-        <label for="{{$id}}" class="col-sm-2 text-wrap mt-2">
+        <label for="{{$id}}" class="col-sm-2 text-wrap mt-2 form-label">
             {{$title}}
 
             @includeWhen(isset($popover),'platform::partials.fields.popover',[